fix: 修复状态机设计缺陷 - 并发安全/SSE/错误恢复
C1: isFailed 推导增加条件（无 generating 锁时才判定），失败时追加 retry action
C2: 新增 POST /panels/retry-failed 端点 + PanelProductionService.retryAllFailedVideos()
C3: checkAndAdvanceProjectState 加 isGenerating 幂等检查，防止并发重复触发
C4: 视频生成超时/异常/失败时写 Redis error + 发 SSE failure 事件
C5: 回滚 Action 加 guardNotGenerating 检查，阻止有异步任务时回滚
C6: 单角色图片生成方法加 progressService.clearError()
M8: PanelProductionService 接入 publishEpisodeScriptDone/publishEpisodeStoryboardDone
M11: GridImageService 接入 publishEpisodeGridStatus（generating/generated/failed）

+++ backend/com/comic/src/main/java/com/comic/statemachine/service/StateChangeEventPublisher.java
@@ -87,7 +87,7 @@ public class StateChangeEventPublisher {
         publishToRedis(projectId, "episode:script_done", payload);
     }
 
-    public void publishEpisodeStoryboardDone(String projectId, int episodeId, int episodeNum, int shotsCount) {
+    public void publishEpisodeStoryboardDone(String projectId, Long episodeId, int episodeNum, int shotsCount) {
         Map<String, Object> payload = new HashMap<>();
         payload.put("episodeId", episodeId);
         payload.put("episodeNum", episodeNum);
@@ -95,7 +95,7 @@ public class StateChangeEventPublisher {
         publishToRedis(projectId, "episode:storyboard_done", payload);
     }
 
-    public void publishEpisodeGridStatus(String projectId, int episodeId, int episodeNum, String gridStatus) {
+    public void publishEpisodeGridStatus(String projectId, Long episodeId, int episodeNum, String gridStatus) {
         Map<String, Object> payload = new HashMap<>();
         payload.put("episodeId", episodeId);
         payload.put("episodeNum", episodeNum);
